feat(types): remove int/uint aliases; suggest explicit width on unknown type
Remove `int`, `Int`, and `uint` from PRIMITIVE_ALIASES in hew-types.
Remove the `"int"` arm from the HIR lowering direct spelling map.
Wire aliases in hew-wirecodec update automatically via `aliases_for`.

When a program uses `int` or `uint` as a type annotation, the checker
now emits a diagnostic with an explicit suggestion:
  unknown type `int`; use `i64` for fixed 64-bit integers or `isize`
  for pointer-sized integers

`UserFacingTy` now displays `i64`/`f64` instead of the former `int`/`float`
aliases so error messages reference only valid type spellings.

Migrate all embedded Hew source in Rust tests and inline stdlib snippets
(MonitorRef, PanicInfo) from `int`/`uint` to `i64`/`u64`.

Four lib-test failures remain: stdlib .hew files on disk still use `int`
(std/process.hew, std/misc/log/log.hew, and friends); these are resolved
in the stdlib migration commit.

+++ hew-hir/tests/actor_call_surface.rs
@@ -126,17 +126,17 @@ fn actor_spawn_send_and_ask_lower_to_explicit_hir_surface() {
     let output = lower_checked(
         r"
         actor Counter {
-            let count: int;
+            let count: i64;
 
-            receive fn increment(n: int) {
+            receive fn increment(n: i64) {
             }
 
-            receive fn print_total() -> int {
+            receive fn print_total() -> i64 {
                 return 1;
             }
         }
 
-        fn main() -> int {
+        fn main() -> i64 {
             let c = spawn Counter(count: 0);
             c.increment(10);
             await c.print_total();
